OFED: Enable installing man pages only if rst2man exists
There is be a problem with the .in.rst documents:
The hash that is used for creating the cached copy (used for systems
that have no rst2man available) is taken from the generated .rst
document, rather than from the original .in.rst document.

For now: enable installing them only if rst2man exists.

Issue: 2582616
Change-Id: I63111c9af79d40c0aa7021c11d7913b282b22c28
Signed-off-by: Mohammad Kabat <mohammadkab@nvidia.com>