Add ENABLE_SRP_COMPAT build option
The srp_daemon.service and srp_daemon_port@.service unit-files
have the parameters are missed in old systemd versions.
This causes the numerous error messages in the logs and can
confuse a user. The parameters restrict privileges that are not
used by the srp_daemon. They are optional and can be skipped. The
ENABLE_SRP_COMPAT build option allows to avoid using of those
parameters. The ENABLE_SRP_COMPAT is enabled for SLES < 15 and
RHEL < 8.

Issue: 2060396
Change-Id: I9e1b0f8b49f91107f877599ef26a29593ea83178
OFED-Status: keep
Signed-off-by: Sergey Gorenko <sergeygo@mellanox.com>