debian: Provide "libibmad-devel" by libibmad-dev
Some of the legacy mlnx libraries use the suffix "devel" instead of
"dev" even for deb development headers. This is a violation of the
Debian policy and should be fixed.

However for now have the libibmad-dev package Provide the old name.

This commit fixes breakage of multiple MLNX_OFED packages that depended
on "libibmad-devel" on Debian and Ubuntu (such as ibsim).

Issue: 1883062
Change-Id: I285019f2325eb9b198c05d916dc8b190bd89025a
Signed-off-by: Tzafrir Cohen <tzafrirc@mellanox.com>
Signed-off-by: Feras Daoud <ferasda@mellanox.com>