rdma-core.spec: use existing %cmake_install and %make_jobs
In SLES, the distribution's %cmake macro runs the build in the
subdirectory build/ . This works because %cmake_install also chdirs
there. However the build of rdma-core fails because we use only %cmake
from the distribution.

While we're at it, also use %make_jobs if set.

And don't try to build with ninja on Fedora.

Note that the build macros on SLES will build with ninja if it is
installed.

This cannot be upstream, as it is about SLES, and Upstream uses a
separate SPEC file for it.

Issue: 1824760
Change-Id: I91bf28289a74add5a3bd0b5397be2bba91fdfe0e
Signed-off-by: Tzafrir Cohen <tzafrirc@mellanox.com>
Signed-off-by: Feras Daoud <ferasda@mellanox.com>