Don't hardcode path of libexec: explicit edit
There are several scripts that assume the directory structure of RedHat
with /usr/libexec . This is not the case with SLES systems, for instance.

The right fix would probably have been to let cmake fix those scripts at
build time. But even upstream's suse spec uses sed instead, and so will
we: if libexec is not the hardwired RedHat value: fix it.

Note that the rdma service is still not enabled by default. This follows
Upstream. The script has an [Install] section but is indeed never
explicitly enabled, and requires a reboot or daemon-reload to be
available.

Issue: 2046224
Change-Id: I5aa1978961731116f0489407a8c8e12235fec45a
Signed-off-by: Tzafrir Cohen <tzafrirc@mellanox.com>