srptools: Don't start srp_daemon on MOFED installation
The MOFED installer assumes that kernel modules are reloaded by the
openibd service when all packages are installed. However, the srptools
post-install script starts the srp_daemon service that depends on
openibd. So, it triggers starting the openibd service in the middle of
MOFED installation, which is not a good idea and can cause various
erros. That's why we have the commit ca7234022862 ("srptools: don't fail
deb install if service fails") to avoid failing installation, and it
works in most cases. But sometimes, starting the openibd service in the
middle of MOFED installation causes critical errors and breaks the
installation.

Starting srp_daemon in the middle of MOFED installation is a bug that
came when MOFED switched to using rdma-core.

Given the above, let's remove starting srp_daemon from the post-install
script. We should also reconsider if the mentioned above commit is still
needed after this change.

Issue: 3470517
Change-Id: If5d6480857e79254c2962da4b0282bd866e63323
Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>