srptools: Do not enable srp_daemon on install
Most MOFED users do not need srp_daemon because it is only needed to
support the legacy storage protocol. However, MOFED installs srp_daemon
and even enables it by default on Debian-based OSes. It is a bad
practice and causes issues. For example, srp_daemon is aimed to work in
small IB fabrics with up to tens of nodes. Srp_daemon generates
significant SA traffic and can even cause SM outage if it is running in
an IB fabric with thousands of nodes. So, many users have to disable
srp_daemon manually. That's why the current default does not look
reasonable. Let's change it.

This commit is irrelevant for upstream because srp_daemon is distributed
as a single package, not a part of the software bundle in this case.
Enabling on install looks reasonable if a user explicitly installs the
srp_daemon package.

Issue: 3613899
Change-Id: I76d24fbf9bad9e289060b4d4fcc4059bf5c89ffe
Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>