mlx5: DR, changing QP creation ts_format to default
When opening a QP in SW steering, one of the attributes is TS
(timestamp) format.
There is a cap that can be queried from the device, but it resides under
RoCe caps table. The problem arises when the RoCe feature is disabled,
and the cap is not read, which leads to pass 0 as the value of the
qp_ts_format. When the system is configured to run with REAL_TIME_TS
format, the creation of the qp will fail.
To overcome this issue, we can use DEFAULT_TS, which will return the
default that is selected by the device.

Issue: 3731634
Signed-off-by: Itamar Gozlan <igozlan@nvidia.com>
Change-Id: If03b1c39f236a8926fbf006a3b7e7675c1199660