mlx5: DR, Using sq ts format when roce disabled
When creating a QP in SW steering, one of the attributes is TS
format (timestamp).
In some devices, we have a limitation that all QP's should have the same
ts_format.
The qp_ts_format cap resides under the RoCE caps table, and the cap is not
queried when RoCe is disabled.
Consider the case when the system is configured to run with REAL_TIME_TS
format (0x2), some QPs are created, and we disable RoCe.
When trying to construct a new QP, we can't use the qp_ts_format, that is
queried from the RoCE caps table.
That leads to passing 0x0 (FREE_RUNNING_TS) as the value of the
qp_ts_format, which is different than the ts_format of the previous
QPs REAL_TIME_TS format (0x2).
To resolve this, we will read the sq_ts_format, which also reflect
the supported ts format for the QP when RoCe is disabled, and override it
with qp_ts_format from the RoCE caps table if RoCe is enabled.

Signed-off-by: Itamar Gozlan <igozlan@nvidia.com>
Issue: 3731634
Change-Id: Iaf755f00711926cdb9aa0beff14fbf8e9b15cbc6