Do build mlx4 and efa providers
Enable building two extra providers: mlx4 and efa.
They are available as libraries, but the configuration and symlink needed
to load it from libibverbs to support their software are removed:
/etc/libibverbs.d/efa.driver
/etc/libibverbs.d/mlx4.driver
$libdir/libibverbs/libefa-rdmav*.so
$libdir/libibverbs/libmlx4-rdmav*.so

Therefore it will be usable for userspace programs that happen to link
with libefa.so.1 or libmlx4.so.1, but not for direct hardware support, and
it will not load automatically.

Issue: 3828530
Issue: 3814955
Change-Id: Ib8db4392b80be38a6f9e0f9a6802c8fdc6fff49b
Signed-off-by: Tzafrir Cohen <tzafrirc@nvidia.com>